Tweak devil collision resolution
Before, devils would get stuck together and jiggle around instead of
separating. This same bug has been coming up in different forms since
we added collision resolution. This time, it was because of rounding
errors when trying to move the devils distances smaller than one pixel.
So I just switched it to use .incrx() and .incry() to allow for
subpixel movement.

Also changed the randomdirection() utility function to push out by
a constant distance instead of a random amount. (I think Luke's
code was already doing this.) This creates a more orderly look.